circuits: zk-circuits: valid-commitments: Add initial test suite
